Remove (unnecessary??) code to explicitly copy over temp .msc file
I think this is already handled by the generic code further down that copies over various auxiliary files from the temp ROM to the output ROM. I also have no idea why it was inside the graphics import part of the function, I think that could actually cause some "bugs" where a later part of the build process fails, but the MSC file of the failed build process makes it in anyway, because it was already copied over. Very weird, never noticed it before.