Fix temp ROM being copied to output ROM and then promptly renamed to it
I believe it used to be the case that the temp ROM was first copied to the output ROM and afterwards "moved" to the output ROM immediately in the loop that is supposed to move over all the auxiliary temp files.

I believe this may have caused potential issues for users with syncing systems, which may pick up on the file change after the copy and then read the output ROM in order to sync it, briefly taking possession of the ROM and stopping the subsequent move from succeeding.

I fixed this by just not copying the temp ROM before the loop, since it's moved in the loop anyway.