Fix auxiliary (and ROM) temp renaming loop ignoring `temp` config var
I believe the loop that renames temporary files to the build output after a successful build used to not take the `temp` config var into account and would just assume it to be "temp" no matter what (hardcoded search string for temporary files was `temp*`). 

I hope to have fixed this by extracting the chosen temp name from the `Config.TempPath` and using a `temp_name + "*"` search for the temp files instead.