Implement comparing old and new tool subgraphs
I had to make some changes, specifically to how library/routine vertices are stored, since unlike for any other files, their file names actually matter, since they're directly converted to macros/labels and also not referenced in any sort of list file by name.

For this purpose, the names of those files are now also stored in the build report and checked for changes.

I'm not terribly happy with the code of the `DependencyGraphAnalyzer`, but it does seem to be working well, so I'm keeping it for now.

Note that as soon as the analyzer finds a modified, missing or arbitrary file vertex in either the old or new subgraph, it returns immediately. The whole subgraph is only explored if they're identical or the very last evaluated vertex is a modified/missing/arbitrary one.