Handle asar `table` command dependencies
I sorta forgot this was a thing, but you can specify a file to use for mapping characters to bytes for tables in asar, which is also a dependency. 

This is now handled by scanning for incsrc/incbin and table commands in the asar resolver. Table files are always treated as binary files, since I believe they cannot introduce any further dependencies.