Potentially fix a potential file access timing bug
In rare cases, it was possible for Lunar Helper to fail to rename a temp file to the final output file. I'm not sure what was (or is?) causing this, but my theory is that if the ROM is open in Lunar Magic and Lunar Monitor tries to reload the ROM, it may have been the case that not all auxiliary files had been renamed yet, since Lunar Helper used to write the build report before renaming the temp files. Since Lunar Monitor reloads the ROM as soon as the build report changes, my theory is that Lunar Magic could in some cases read auxiliary files at the same time that Lunar Helper tries to write to them, causing a file access denied error.

I cannot confidently say whether this is actually the case, but I'm struggling to come up with other reasons why auxiliary or ROM files would be accessed at this precise timing, so I hope simply renaming the files *before* writing the build report is an adequate fix.