Prompt user before build if there may be unexported resources in ROM
Also adds a `-v` or `--volatile-resource-pref` command line option when building, quick building or packaging from the command line which can be used to specify what to do if the build runs into potentially unexported resources in a pre-existing ROM.

`Y` will export all resources and then continue building, `N` will *not* export resources and just continue building and `C` will error out and cancel the build (`C` is the default).

ROMs built by Lunar Helper will be marked as non-volatile, meaning more or less the only practical way to have the ROM be marked as volatile (aka containing potentially unexported resources) is to edit it with an uninjected Lunar Magic. Should mean that if you've encountered the warning once, you will never see it again unless you accidentally or intentionally use an uninjected Lunar Magic on the ROM.