Add config vars for dependency lists
Adds `build_order` and `quick_build_triggers` config variables.

These can be used like this:
```
build_order
[
    PIXI
    Levels
    Map16
    Patches/some_random_patch.asm
    GPS
    Patches
    Graphics
    (etc.)
]
```
and
```
quick_build_triggers
[
     PIXI -> AMK, UberASM, Patches
     Map16 -> GPS, Levels
     Patches/some_random_patch.asm -> GPS, UberASM
     (etc.)
]
```
`build_order` simply specifies in which order resources should be inserted. The idea is probably going to be to specify every tool/resource that is actually going to be inserted in there. 
`Patches` is a shorthand for all patches that were not otherwise mentioned in the build order.

`quick_build_triggers` is similar, though with a different format. It lets you specify which tools/resources should be reinserted if a change is detected in another tool/resource. I.e. `PIXI -> AMK, UberASM, Patches` means "if a file in PIXI changes, reinsert AMK, UberASM and all patches in that order". Note that `Patches` will probably mean literally all patches, not just the ones that aren't otherwise mentioned in this list.

Note that only the actual parsing of these lists is (hopefully) implemented at this point. Actual functionality is coming later.