Implement bulk of dynamic build order and quick build trigger code
This should hopefully mostly work. It's a quite drastic rework of when stuff like paths to tools are checked and the general mechanism for inserting resources behind the scenes.

The basic idea of this is to enable users to customize the build order as well as customizing which tools/resources should get reinserted by Quick Build when a different resource has changed.

This is an example configuration for the `build_order` variable:
```build_order
[
	Graphics
	Map16
	Patches/message_box_expansion.asm
	GlobalData
	Patches
	PIXI
	Levels
	PIXI
	AddMusicK
	UberASM
	SharedPalettes
]```

The resources listed will be inserted in this order during the Build process. Note that resources can be listed more than once, this mostly so that PIXI can be inserted twice if need be for LM versions 3.31 and higher.

Note that if a resource is listed but not specified Lunar Helper will throw an error to let the user know. I.e. if `pixi_path` were not specified, this particular configuration would throw an error since PIXI is in the build order but cannot actually be used.

The supported names in this list are any paths to individual patches, as well as Pixi, AddMusicK, Gps, UberAsm, Graphics, ExGraphics, Map16, SharedPalettes, GlobalData, TitleMoves, Levels and Patches (note that Patches is a shorthand for all other patches that are not explicitly mentioned by name in the `build_order` but are listed in the `patches` list).

This is an example of the `quick_build_triggers` variable:

```quick_build_triggers
[
	UberASM -> AddMusicK
	UberASM -> Levels
	Levels -> AddMusicK
	UberASM -> Patches\message_box_expansion.asm
	Patches/message_box_expansion.asm -> PIXI
]```

The syntax for this is `X -> Y`, meaning "if X is re-inserted, also re-insert Y". I.e. the first line in this particular list means "if Quick Build re-inserts UberASM, also reinsert AddMusicK".

Note that these triggers cascade, i.e. UberASM being re-inserted will also result in PIXI being re-inserted, since line 4 will trigger the re-insertion of the patch, which will then trigger re-insertion of PIXI due to line 5.

Lunar Helper will automatically ensure that as few insertions as possible are performed, i.e. if UberASM is re-inserted, the resulting insertion order will be UberASM -> Levels -> AddMusicK, not UberASM -> AddMusicK -> Levels -> AddMusicK.

Note that Lunar Helper does not make any guarantees about when exactly the resource on the right side of a rule will be inserted. The only guarantee for made by a rule of the form `X -> Y` is that if X is re-inserted, Y will be reinserted *at some point* after X.

Note that `Patches` can only be used in `build_order`, not in `quick_build_triggers`.