Simplify default config files
Making an effort to remove config variables that aren't crucial from the config files in an attempt to make configuration more approachable. The idea is that most people won't need any of the more advanced config variables and if they do they can read up on them in the wiki and then add them themselves.