Rework `.lunar_helper` location and (Ex)Graphics ex-/import consistency
Exporting and importing (Ex)Graphics should now work consistently, even if `output` and `temp` ROMs are in different locations and/or not in the project root folder.

The `.lunar_helper` directory should now also always be generated in the folder of the `output` ROM instead of the project root for better compatibility with Lunar Monitor, which expects it to be in this location.