Fix weird behavior when a space is inserted at end of ROM name in LM
For some reason, it seems like sometimes a space is getting inserted after the ROM name in Lunar Magic's internal memory. Apparently, this space then made it to the FLIPS command line call when Lunar Monitor exported global data, which would then cause FLIPS to interpret the extension as `.smc ` (note the space), which is different from `.smc`, which thus caused the header to be part of the BPS, which would cause problems. 

Very strange, not sure what causes the presence or absence of the space in Lunar Magic's memory, but the change seems to have fixed it (for now).