Skip message box prompts when Lunar Helper is invoked with CLI arguments
This is so that message box prompts won't halt automated scripts forever. Probably would be more robust to offer a command line argument for configuring whether prompts should halt or not, just so that people who prefer using the command line interface directly can still answer prompts instead of being required to launch Lunar Helper's menu to do this, but for now this will do as standard behavior, I think.