Ensure LML doesn't inject twice
This was causing injection to sometimes not work, because LML would stall indefinitely on the second load, since it would wait for a pipe read that would never come, because LM was already injected and would not read from the new new pipe.

Now when injecting into a running Lunar Magic, I check whether LM is already injected and then don't inject again, which should get rid of the indefinite stall.