Remove lag when Lunar Monitor exports resources during saves
I think this should work fine, there might be some problems if two saves occur within an extremely short timeframe since the first export task could still be reading the ROM while the second one tries to access it.

Not launching the "Export all" button function in a thread cause it takes a lot longer and conflicts are probably more likely if I just let you keep doing stuff in that case. You really shouldn't need to use that button all that much anyway so probably not a big deal.